Refactored Codebase
Created new Entity and Paddle classes. Entity is a generic entity that is drawn on the screen with a position vec3/transform matrix and a window to be drawn to. Paddle is a derived class of Entity with and update function for moving in response to specified keypresses.

Updated the drawing logic to use more modern OpenGL features such as glCreateBuffers(), glNamedBufferdata(), glVertexAttribFormat(), glVertexAttribBinding(), and glDrawElementsBaseVertex(). These new functions required me to update the GLFWwindow hint and get new versions of the glad headers and source file.

Also updated the vertex shader to take in an MVP (Model, View, Projection) matrix via a uniform. This is a common matrix used in most 3D games for transforming vertex coordinates from model space to world space to camera/view/eye space to clip space. Right now, since this is a simple 2D application, the View and Projection matrices are just the identity matrix.

Also also, added glm to the project via include.